Create layout and navbar components, set basic styling
Created a layout component displaying the navbar component and the main
content of the page depending on the route. The routing is done using the
react-router module. If you are reading this I have exams starting on 5th
of January and having to study and complete this project is kinda stressful.